refactor: 레포지토리 추상화
- 모든 Repository 클래스들을 추상화
- 이들의 빈 등록을 Configuration에서 관리함으로써 레포지토리 교체의
  편의성 향상
- 순환 참조 문제를 해결하기 위해 인터셉터에서 AuthService가 아닌
  AuthRepository를 사용하도록 변경
